feat: redo 7 with array instead of linked-list
